Initial work on a schema for the osvr_server config file.
This starts work on #5 - external devices is basically done,
the special server options section is done, and the plugins and
drivers sections are mostly done. Notably, aliases (and routes)
aren't done, and those will be using a shared schema. In any case,
this should be more useful than not having a schema at all.